wikipedia_langraph_chatbot/requirements.txt
langgraph==0.6.7
langchain-google-genai==2.1.10
langchain-core==0.3.76
python-dotenv==1.1.1
requests==2.32.5
typing-extensions==4.15.0
